fix: segfault in ex01, reverted ex02

+++ cpp09/ex01/includes/RPN.hpp
@@ -34,7 +34,7 @@ class RPN
 		int _checkinput();
 
 		std::string 	_input;
-		std::stack<int>	_stack;
+		std::stack<long int>	_stack;
 };
 
 #endif
